LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/script-text-modifications-expected.txt


FAIL Modify inline script element's text after prepare-a-script before evaluation assert_equals: <script>'s innerText should be already modified expected "t.unreached_func(\"This should not be evaluated\")();" but got "\nt.step(() => {\n    // When this is evaluated after the stylesheet is loaded,\n    // script0's innerText is modified by the async script above,\n    // but the evaluated script is still the original script here,\n    // not what is overwritten, because \"child text content\" is taken in\n    // #prepare-a-script and passed to \"creating a classic script\".\n    var s = document.getElementById('script0');\n    assert_equals(s.innerText,\n                  't.unreached_func(\"This should not be evaluated\")();',\n                  \"<script>'s innerText should be already modified\");\n    assert_equals(s.text,\n                  't.unreached_func(\"This should not be evaluated\")();',\n                  \"<script>'s text should be already modified\");\n    t.done();\n  });\n"
